+++ src/bp/ui-admin/src/Pages/Workspace/Bots/index.tsx
@@ -19,7 +19,7 @@ import React, { Component, Fragment } from 'react'
 import { connect } from 'react-redux'
 import { generatePath, RouteComponentProps } from 'react-router'
 import { Alert, Col, Row } from 'reactstrap'
-import { setupServerPackageLoop } from '~/telemetry/middleware'
+import { setupServerPackageLoop } from '~/telemetry/offlineServerFallback'
 import { toastSuccess } from '~/utils/toaster'
 import { toastFailure } from '~/utils/toaster'
 import { filterList } from '~/utils/util'
