+++ src/bp/ui-admin/src/telemetry/telemetryType.d.ts
@@ -1,0 +1,13 @@
+export interface StoreInfoType {
+  [key: string]: Function
+}
+
+export interface EventPackageInfoType {
+  [key: string]: EventPackageInfoBody
+}
+
+export interface EventPackageInfoBody {
+  locked: boolean
+  timeout: string
+  getPackage: Function
+}
